move files to a flat structure
